Implemented system of backup for quotas for yt
A reworked YouTube query and YouTube metadata to use convars for YouTube API to use as replacement (if people want it for some reason), to replace the api key for other one, and use a system as backup with other convar.
Moved PlayX initial variable to be on the loader, so can I add version and updated version variables.
"PlayX.Version" and "PlayX.VersionUpdated" so that is now passed down to server side AND client-side as variables "playx_version" and "playx_version_updated".
I had it put on panel on settings but didn't liked the way it looked, and got a bit cluttered so I removed. That variables does NOT do any thing, it is just for show/read only.
These next convars does work, you can change it, or just forget it, it will still work as intended:
The first key for api is a convar called "playx_google_api_v3_youtube_key_main"
The second key for api convar is called "playx_google_api_v3_youtube_key_backup"
This system of backup could not be tested properlly because we are not out of quota yet, but once we hit it, the system should work greatly to prevent out of quota errors like it had some months ago.
Replaced some functions that should not be called by gmod (they don't like it) to some proper ones "GetConvarString" replaced by "GetConvar("variable"):GetString() its a pain but should prevent some depreciation warnings in the future.