Removed by sk89q from the logo
I decided to remove the original owner name from logo, since be TOTALLY abandoned the project for 12 years now.
I do must of the work now. So there is no need to put the original owner as image for a project that does not have is input in years.

I really liked the new logo, is more clean now